modified session manager
session manager also has the engine log inside the struct, which then will pass it down to other sessions that are added into the manager

this makes it so that new session does not need the log parameter. It still has the variable in the struct, but it gets passed from the manager now.